Using a custom pico board so that the correct peripherals and flash autorun is enabled
The many pico custom designs use  different peripherals, different pin outs and different flash memory from the standard pico board.
These boards have a library in the sdk and this repository shows how to configure a source to use these libraries and to compile the code for the correct board, allowing for the correct hardware to be used, and for code to ported between boards and for boards to auto run the executable on reset.